requirements.txt
pyrebase
firebase_admin
requests
python-dotenv